+++ src/main/kotlin/Transformation.kt
@@ -217,7 +217,7 @@ data class Transformation(var matrix:HomMatrix= HomMatrix(ID.copyOf()), var invm
     {
         return Transformation(
             matrix=this.matrix*other.matrix,
-            invmatrix=other.invmatrix*this.matrix
+            invmatrix=other.invmatrix*this.invmatrix
         )
     }